refactor: delete wigner3j for mul + add features in ylm
